tlf_journal: add single op mode
Single op mode allows an application to tell the journal not to flush
any blocks until after the application has signalled its work is
finished.  It may flush blocks in the meantime.  Then, the MDs and
remaining blocks are squashed into a single revision and put to the
server, so all the changes in the single op show up atomically to
other TLF readers.

Issue: KBFS-2379